fix: keep revert rpc in default unsafe admin api
