+++ madara/node/src/cli/rpc.rs
@@ -91,8 +91,9 @@ pub struct RpcParams {
     #[arg(env = "MADARA_RPC_ADMIN_EXTERNAL", long, default_value_t = false)]
     pub rpc_admin_external: bool,
 
-    /// Enables unsafe admin RPC methods. When built with the `replay` Cargo feature,
-    /// this also includes replay-specific methods that can modify blockchain state.
+    /// Enables unsafe admin RPC methods such as `revertToAndShutdown`. When built
+    /// with the `replay` Cargo feature, this also includes replay-specific methods
+    /// such as `setCustomBlockHeader` that can modify blockchain state.
     /// Use with extreme caution. Requires `--rpc-admin` to be enabled.
     #[arg(env = "MADARA_RPC_UNSAFE", long, default_value_t = false, requires = "rpc_admin")]
     pub rpc_unsafe: bool,
